+++ packages/ui/src/icons/index.ts
@@ -60,3 +60,5 @@ export * from './LogoRampCaptionIcon';
 export * from './LogoRampIcon';
 export * from './DotArrowRightIcon';
 export * from './TourIcon';
+export * from './AddIcon';
+export * from './RemoveIcon';